requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
schedule==0.6